Source hash: 84f3e3d9f2f1e1a7a3db3e3b630fbd507c607fe7
Remaining commits: 0
Merge conflict resolutions:
https://quexington.github.io/merge-diffs/bea085d265798537b15d99525ee4f61d12883ee9.html

+++ chia/types/unfinished_header_block.py
@@ -34,4 +34,4 @@ class UnfinishedHeaderBlock(Streamable):
 
     @property
     def total_iters(self) -> uint128:
-        return self.reward_chain_block.total_iters
+        return uint128(self.reward_chain_block.total_iters)
